+++ NEWS.md
@@ -7,6 +7,8 @@ dates to returned data (#24).
 - `get_dataset()` will now suggest multiple dataset names, when the dataset 
 you've asked for doesn't exist (i.e. there's a typo) and there are multiple 
 likely candidates  (#28).
+- Two new functions `list_datasets()` and `list_resources()` allow browsing
+available datasets and resources (#10).
 
 # phsopendata 0.1.0 (2021-07-22)